# Benchmark encode and cut an encoding inefficiency.