--- a/test/codec_perf/perf.proto
+++ b/test/codec_perf/perf.proto
@@ -1,0 +1,11 @@
+// Messages used for performance tests that may not reference grpc directly for
+// reasons of import cycles.
+syntax = "proto2";
+
+package codec.perf;
+
+// Buffer is a message that contains a body of bytes that is used to exercise
+// encoding and decoding overheads.
+message Buffer {
+  optional bytes body = 1;
+}