@iamqizhao, this pull request is apropos of the private conversation
between us, wherein you stated the desire to start adding elementary
benchmarks into the platform as next-step.  If we are using Gerrit
(already), please let me know the project root, and I will redirect this
accordingly:

This commit introduces the first microbenchmark for grpc, wherein
`encode` is benchmarked according to message size.  A conclusion of
the benchmark is that the removal of type switching found in
`binary.Write`, which is used in `encode` produces the following
encoding time and memory allocation footprint:

```
$ # Return to previous commit but benchmark.
$ go test ./... -test.bench="Benchmark*" > /tmp/before
$ # Return to working copy.
$ go test ./... -test.bench="Benchmark*" > /tmp/after
$ benchcmp /tmp/before /tmp/after
benchmark               old ns/op     new ns/op     delta
BenchmarkEncode1        1451          942           -35.08%
BenchmarkEncode10       1466          1067          -27.22%
BenchmarkEncode100      1966          1648          -16.17%
BenchmarkEncode1000     4390          3910          -10.93%

benchmark               old MB/s     new MB/s     speedup
BenchmarkEncode1        5.51         8.49         1.54x
BenchmarkEncode10       11.60        15.93        1.37x
BenchmarkEncode100      54.40        64.92        1.19x
BenchmarkEncode1000     229.59       257.77       1.12x

benchmark               old allocs     new allocs     delta
BenchmarkEncode1        6              3              -50.00%
BenchmarkEncode10       7              4              -42.86%
BenchmarkEncode100      8              5              -37.50%
BenchmarkEncode1000     8              5              -37.50%

benchmark               old bytes     new bytes     delta
BenchmarkEncode1        384           328           -14.58%
BenchmarkEncode10       400           344           -14.00%
BenchmarkEncode100      736           680           -7.61%
BenchmarkEncode1000     2560          2504          -2.19%
```

..., which is apropos of the comment in [encoding/binary]
(http://golang.org/pkg/encoding/binary), wherein ...

> This package favors simplicity over efficiency.

... is stated.

If `encode` is deemed to need further memory efficiencies, a mechanism
whereby a `proto.Buffer` is retained may be warranted, which is why the
original TODO remains.  The proposed improvement in this change is
simple and low-hanging.

I did not want to introduce yet-another protocol buffer message for
tests, but the ones under ...

> interop/grpc_testing/test.proto
> test/grpc_testing/test.proto

... have a fundamental dependency on `grpc` package due to their
generated stubs, which produces a cycle in the imports if the benchmark
were to attempt to import them for profiling.  The newly created ...

> test/grpc_message/test.proto

... protocol buffer package has no generated RPC service stubs, which
means it can be imported into the `grpc` package root without cycle.