@iamqizhao 2015-03-02
Hi Matt,

Thanks for making this happen. Can you move the benchmark to a separate test file such as microbenchmark_test.go?