@iamqizhao 2015-03-02
I would also suggest moving it to test/ dir. One positive side effect is that you can reuse the proto (and its generated code) there.