@matttproud 2015-03-02
I have gone ahead and renamed it.  As for moving it, I am curious for your suggestions:

We are benchmarking an unexported, internal function `encode` at the moment, which belongs to the top-level package.  There will be more that we benchmark, but this is just the beginning.

The only way we can benchmark it while moving it to another package appears to be …

1. Exporting the function (highly undesirable due to public API cluttering).
2. Moving it to an [internal package](http://golang.org/s/go14internal), provided that meets the Golang versioning compatibility requirements for grpc and is _actually implemented_ in upsteam Go (a quick search at stdlib in 1.4.1 answers: no).

What are your thoughts?