@iamqizhao 2015-03-02
Hi Matt,

I just realized you only benchmarked encode function instead of the
end-to-end thing (sorry, I only glanced the file names and did not look
into your code last night.). If that is the case and you think it has some
value to do, your original choice is better here. Actually, I did not plan
to do this kind of microbenchmark. But it seems it is good to have if you
already started working on it. Thanks.

On Mon, Mar 2, 2015 at 10:08 AM, Matt T. Proud <notifications@github.com>
wrote:

> I have gone ahead and renamed it. As for moving it, I am curious for your
> suggestions:
>
> We are benchmarking an unexported, internal function encode at the
> moment, which belongs to the top-level package. There will be more that we
> benchmark, but this is just the beginning.
>
> The only way we can benchmark it while moving it to another package
> appears to be …
>
>    1. Exporting the function (highly undesirable due to public API
>    cluttering).
>    2. Moving it to an internal package <http://golang.org/s/go14internal>,
>    provided that meets the Golang versioning compatibility requirements for
>    grpc and is *actually implemented* in upsteam Go (a quick search at
>    stdlib in 1.4.1 answers: no).
>
> What are your thoughts?
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/grpc/grpc-go/pull/87#issuecomment-76765389>.
>