@matttproud 2015-03-02
Yeah, there is value in instrumenting these smaller units to get a better approximation of the whole.  I started with`encode`, because of your TODO in the function itself.  More broadly, I have a few ideas about end-to-end benchmarking.  I just wanted to start with the simpler cases and expand from there.