@iamqizhao 2015-03-02
On Mon, Mar 2, 2015 at 12:03 PM, Matt T. Proud <notifications@github.com>
wrote:

> Yeah, there is value in instrumenting these smaller units to get a better
> approximation of the whole. I started withencode, because of your TODO in
> the function itself. More broadly, I have a few ideas about end-to-end
> benchmarking. I just wanted to start with the simpler cases and expand from
> there.
>
sgtm. then either rename microbenchmark_test.go to
unit_microbenchmark_test.go or put it back into rpc_util_test.go as you
originally did. Thanks.

> —
> Reply to this email directly or view it on GitHub
> <https://github.com/grpc/grpc-go/pull/87#issuecomment-76802166>.
>