@matttproud 2015-03-02
Returned to rpc_util_test.go.