@iamqizhao 2015-03-02
Probably rename the new proto to codec_perf and put it under test/codec_perf correspondingly. We could enhance it to benchmark various codec related unit benchmark. The current name is a bit misleading in test/ dir.